pyproject.toml
[project]
name = "GrANv1.0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
      "pandas", "numpy", "matplotlib", "jax", "dm-haiku", "optax", "scikit-learn", "einops", "scipy", "tqdm", "statsmodels", "seaborn",
      "flax",
]
